// web/src/lib/api-types.ts
export type Role = "USER" | "ADMIN";

export interface ApiUser {
  id: string;
  email: string;
  name: string | null;
  avatarUrl: string | null;
  role: Role;
  createdAt: string;
  updatedAt: string;
}